# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Babussalam Mukholif API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthStatus {
  authenticated: boolean;
  username: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export interface Violation {
  id: number;
  studentName: string;
  kelas: string;
  asrama: string;
  waktu: string;
  jenisPelanggaran: string;
  /** @nullable */
  catatan?: string | null;
  createdAt: string;
}

export interface ViolationInput {
  studentName: string;
  kelas: string;
  asrama: string;
  waktu: string;
  jenisPelanggaran: string;
  catatan?: string;
}

export interface ViolationUpdate {
  studentName?: string;
  kelas?: string;
  asrama?: string;
  waktu?: string;
  jenisPelanggaran?: string;
  catatan?: string;
}

export interface StudentSummary {
  studentName: string;
  kelas: string;
  asrama: string;
  violationCount: number;
  /** @nullable */
  lastViolation?: string | null;
}

export interface DashboardStats {
  totalViolations: number;
  totalMukholif: number;
  totalKelas: number;
  totalAsrama: number;
  violationsToday?: number;
  violationsThisWeek?: number;
  violationsThisMonth?: number;
}

export interface TopViolator {
  studentName: string;
  kelas: string;
  asrama: string;
  violationCount: number;
}

export interface ViolationTypeCount {
  jenisPelanggaran: string;
  count: number;
}

export interface ViolationClassCount {
  kelas: string;
  count: number;
}

export type ListViolationsParams = {
studentName?: string;
kelas?: string;
asrama?: string;
jenisPelanggaran?: string;
};

export type ListStudentsParams = {
search?: string;
kelas?: string;
asrama?: string;
};

export type GetTopViolatorsParams = {
limit?: number;
};

export type GetRecentViolationsParams = {
limit?: number;
};