# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Babussalam Mukholif API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalViolations: number;
  totalMukholif: number;
  totalKelas: number;
  totalAsrama: number;
  violationsToday?: number;
  violationsThisWeek?: number;
  violationsThisMonth?: number;
}