# lib/api-zod/src/generated/types/topViolator.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Babussalam Mukholif API
 * OpenAPI spec version: 0.1.0
 */

export interface TopViolator {
  studentName: string;
  kelas: string;
  asrama: string;
  violationCount: number;
}