# lib/api-zod/src/generated/types/violationTypeCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Babussalam Mukholif API
 * OpenAPI spec version: 0.1.0
 */

export interface ViolationTypeCount {
  jenisPelanggaran: string;
  count: number;
}